requirements.txt
asgiref==3.5.0
Babel==2.9.1
Django==4.0.3
django-filter==21.1
django-money==2.1.1
django-tables2==2.4.1
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
